refactor : view 정리
